requirements.txt
fastapi
uvicorn
aiomysql
jwt
pandas
google-auth==2.35.0
cryptography==42.0.1
python-jose==3.3.0
python-dotenv>=1.0.1
loguru>=0.7.2
aiosmtplib
bcrypt
jinja2
pydantic[email]
geoip2
requests_oauthlib
fastapi_discord
authlib
httpx
redis==5.0.2
pypika-tortoise==0.1.6
tortoise-orm[asyncmy,asyncpg]==0.20.0
aiokafka
kafka-python
kafka-python-ng
APScheduler==3.11.0
web3==7.6.1
apscheduler
boto3
python-multipart
dbutils
